Expose APIs that are needed
And move the protocol to a `protocol` sub-module.
